Fixed bug that resulted in mass being an integer.

+++ drake/systems/plants/RigidBody.cpp
@@ -205,11 +205,11 @@ void RigidBody::set_contact_points(const Eigen::Matrix3Xd& contact_points) {
   contact_points_ = contact_points;
 }
 
-void RigidBody::set_mass(int mass) {
+void RigidBody::set_mass(double mass) {
   mass_ = mass;
 }
 
-int RigidBody::get_mass() const {
+double RigidBody::get_mass() const {
   return mass_;
 }
